requirements.txt
# Core Libraries
opencv-python
bcrypt
cryptography

# GUI and System Interaction
pystray
Pillow
plyer
pywin32